+++ include/funcoes.h
