+++ src/funcoes.cpp
